Travis CI build configuration
The badge at the top of the README will show the current build status,
and links to the [project page on Travis](https://travis-ci.org/TL1-fa17/projet3)

Running `ant` from the root of the repository should make a clean build
and run the tests. JUnit JARs included for simplicity, they are probably
redundant with the ones used by Eclipse, but I'm too lazy to find a
nicer solution (ideally the whole project should be managed with Maven…)

See #6.